% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01atributuak.R
\name{Atributua}
\alias{Atributua}
\title{Atributua instantzia sortzeko funtzioa}
\usage{
Atributua(izena, balioak, mota = NULL)
}
\arguments{
\item{izena}{Atributuaren izena (karaktere kate bat).}

\item{balioak}{Atributuaren balioak (zenbakizkoak edo kategorikoak).}

\item{mota}{Atributuaren mota (aukerazkoa: `"numeric"` edo `"categoric"`).}
}
\value{
"Atributua" klaseko objektu bat.
}
\description{
Atributu bat sortzen du, bere izena, mota eta balioekin.
}
\examples{
Atributua("adina", c(10, 20, 30))
}
