% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01atributuak.R
\name{DatuMultzoa}
\alias{DatuMultzoa}
\title{DatuMultzoa instantzia sortzeko funtzioa}
\usage{
DatuMultzoa(datuak, klase_izena = NULL)
}
\arguments{
\item{datuak}{Data.frame bat atributuak eta klase aldagaia jasotzen dituena.}

\item{klase_izena}{Klase gisa erabili beharreko zutabearen izena (aukerakoa).}
}
\value{
"DatuMultzoa" klaseko objektu bat.
}
\description{
'DatuMultzoa' klaseko objektu bat sortzen du `data.frame` batetik.
}
\examples{
datuak <- data.frame(adina = c(23, 45, 34), sexua = c("M", "F", "M"))
DatuMultzoa(datuak, klase_izena = "sexua")
}
