% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02diskretizazioa.R
\name{discretizeEF}
\alias{discretizeEF}
\alias{discretizeEF,numeric-method}
\alias{discretizeEF,Atributua-method}
\title{Equal Frequency diskretizazioa}
\usage{
discretizeEF(x, num.bins)

\S4method{discretizeEF}{numeric}(x, num.bins)

\S4method{discretizeEF}{Atributua}(x, num.bins)
}
\arguments{
\item{x}{Diskretizatzeko objektua (`numeric` edo `Atributua`)}

\item{num.bins}{Zenbat tarte sortu nahi diren adierazten duen balio zenbakizkoa}
}
\value{
\itemize{
  \item `numeric` sarreran: diskretizatutako balioak eta ebaki puntuak dituen zerrenda bat.
  \item `Atributua` sarreran: kategoria tarteetan bihurtutako `Atributua` objektua eta bere ebaki puntuak.
}
}
\description{
`discretizeEF()` funtzioak balio jarraituak kuantilen arabera
tarte berdinetan banatzen ditu (maiztasun berdineko binak).
Funtzio honek `numeric` eta `Atributua` klaseetarako metodoak ditu.
}
\examples{
x <- c(10, 20, 25, 30, 40, 50, 60, 70)
discretizeEF(x, num.bins = 3)

}
