% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07plotting.R
\name{Korrelazioak_irudikatu}
\alias{Korrelazioak_irudikatu}
\title{Korrelazioen irudikapena}
\usage{
Korrelazioak_irudikatu(M)
}
\arguments{
\item{M}{Data.frame edo matrize bat, non zutabe guztiak zenbakizkoak diren.}
}
\value{
Korrelazio-matrizearen heatmap bat bistaratzen du.
}
\description{
Dataset baten korrelazio-matrizea edo informazio mutuala irudikatzen du `ggplot2` erabiliz.
Zenbakizko aldagaietarako korrelazioak kalkulatzen dira.
*Adi!!!* Jatorrizko datuak sartu behar dira, funtzioak falkulatzen ditu korrelazioak eta bertatik irudikatzen ditu.
}
\examples{
df <- data.frame(
  adina = c(20, 25, 30, 35, 40, 45, 50),
  soldata = c(5000, 4000, 3000, 2000, 1000, 800, 600),
  altuera = c(1.60, 1.65, 1.70, 1.75, 1.78, 1.82, 1.85)
)
Korrelazioak_irudikatu(df)

}
