% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01atributuak.R
\docType{package}
\name{Permulan-package}
\alias{Permulan}
\alias{Permulan-package}
\title{Permulan: Datuen aurreprozesamendurako eta azterketa estatistikorako tresna multzoa (S4 egiturak erabiliz)}
\description{
Pakete honek datuen aurreprozesamendurako eta azterketa estatistikorako tresna multzoa eskaintzen du, S4 klaseen bidez egituratuta. Hau r-n sortu dudan lehen paketea da.
}
\author{
\strong{Maintainer}: Maddi Olaetxea \email{maddiolausa@gmail.com}

}
\keyword{internal}
