% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01atributuak.R
\name{show,DatuMultzoa-method}
\alias{show,DatuMultzoa-method}
\title{DatuMultzoa objektua erakusteko metodoa}
\usage{
\S4method{show}{DatuMultzoa}(object)
}
\arguments{
\item{object}{DatuMultzoa klaseko objektua.}
}
\description{
`show()` metodoa 'DatuMultzoa' klaseko objektuak pantailaratzeko.
}
