% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07plotting.R
\name{Visualizar_ROC}
\alias{Visualizar_ROC}
\title{ROC kurbaren irudikapena (AUC plot)}
\usage{
Visualizar_ROC(resultado)
}
\arguments{
\item{resultado}{`roc_analisi()` funtzioaren emaitza (`curva` eta `AUC` elementuak dituen zerrenda)}
}
\value{
ROC kurba marrazten du pantailan.
}
\description{
`roc_analisi()` funtzioaren emaitzetatik ROC kurba marrazten du eta
AUC balioa irudikatzen du.
}
\examples{
df <- data.frame(
  adina = c(20, 25, 30, 35, 40, 45, 50),
  soldata = c(5000, 4000, 3000, 2000, 1000, 800, 600),
  altuera = c(1.60, 1.65, 1.70, 1.75, 1.78, 1.82, 1.85)
)
Korrelazioak_irudikatu(df)

}
